Added thermistor map to config (metadata for h5 file).
