+++ Config/thermistors.csv
@@ -1,0 +1,41 @@
+digitizer	 thermistor	 location
+        1	          1	 Cal. target (upper left)
+        1	          2	 Cal. target (lower left)
+        1	          3	 Cal. target (center)
+        1	          4	 Cal. target (bottom left)
+        1	          5	 Cal. target (bottom right)
+        1	          6	 Cal. target (upper right)
+        1	          7	 Cal. target (lower right)
+        1	          8	 Cal. target (top middle)
+        2	          1	 AMR-(ch18/ch24)-internal QV
+        2	          2	 AMR-(ch34)-internal QV
+        2	          3	 AMR-(ch18/ch24)-external QH
+        2	          4	 AMR-(ch34)-external QH
+        2	          5	 AMR-NS-18/24 H
+        2	          6	 AMR-NS-34 H
+        2	          7	 AMR-NS-18/24 V
+        2	          8	 AMR-NS-34 V
+        3	          1	 5V PSU
+        3	          2	 On-board computer
+        3	          3	 Paraboloid (backside bottom)
+        3	          4	 Paraboloid (backside middle)
+        3	          5	 Paraboloid (backside top)
+        3	          6	 Motor controller
+        3	          7	 Motor
+        3	          8	 HyMS IF back-end
+        4	          1	 HyMS IF front-end (location 1)
+        4	          2	 HyMS IF front-end (location 2)
+        4	          3	 HyMS IF front-end (location 3)
+        4	          4	 HyMS IF front-end (location 4)
+        4	          5	 HyMS IF front-end (location 5)
+        4	          6	 HyMS IF front-end (location 6)
+        4	          7	 HyMS IF front-end (location 7)
+        4	          8	 ABEB (external)
+        5	          1	 Optical bench MW-horn
+        5	          2	 HyMS RF front-end (location 8)
+        5	          3	 HyMS RF front-end (location 9)
+        5	          4	 HyMS RF front-end (location 10)
+        5	          5	 IF down-conversion plate (left)
+        5	          6	 IF down-conversion plate (right)
+        5	          7	 Digital back-end (left)
+        5	          8	 Digital back-end (right)
