+++ hammr/masterclient.py
@@ -12,7 +12,7 @@ import time
 from datetime import datetime
 from subprocess import Popen
 
-from create_log import create_log
+from utils import create_log
 from motorcontrol import MotorControl
 from filepaths import ACQ_CONFIGS, ACQ_CONFIGS_TMP, ACQ_DATA, PATH_TO_GENCLIENT, PATH_TO_PYTHON
 from genericparser import GenericParser
